requirements.txt
# Core packages
matplotlib==3.7.1
numpy==1.26.1
opencv-python==4.9.0.80
pyautogui==0.9.53
PyYAML==6.0
Cython==0.29.38

# Deep learning
tensorflow-macos==2.10.0
torch==2.0.1
torchvision==0.16.2

# Utility
PyGetWindow==0.9.0
